move Report component to parent dir
