+++ src/reporter/App.js
@@ -5,7 +5,7 @@ import Tab from 'react-bootstrap/Tab'
 import Row from 'react-bootstrap/Row'
 import Col from 'react-bootstrap/Col'
 import Nav from 'react-bootstrap/Nav'
-import Report from './components/Report'
+import Report from './Report'
 import './App.css'
 import 'bootstrap/dist/css/bootstrap.min.css'
 import results from './results'
